python/requirements.txt
# These are mirrored in setup.py as install_requires,
# which is what the users of the ray package will install. The rest of this file
# sets up all the packages necessary for a /developer/ of Ray.
#
# In short, if you change it here, PLEASE also change it in setup.py.
#
# setup.py install_requires
aiohttp==3.7
aioredis < 2
click >= 7.0
cloudpickle
colorama
colorful
filelock
gpustat
grpcio >= 1.32.0, <= 1.49.1; python_version < '3.10'
grpcio >= 1.42.0, <= 1.49.1; python_version >= '3.10'
jsonschema
msgpack >= 1.0.0, < 2.0.0
numpy >= 1.16
opencensus
prometheus_client >= 0.7.1
protobuf >= 3.8.0, < 4.0.0
py-spy >= 0.2.0
pydantic >= 1.8, < 1.10.0
pyyaml
redis >= 3.5.0, < 4.0.0
requests
virtualenv >= 20.0.21
# For test_cli.py
moto
# For test_joblib.py
sklearn

## setup.py extras
atari_py
dm_tree
flask
gym[atari]
lz4
opencv-python-headless==4.3.0.36
pandas==1.0.5
scipy==1.4.1
tabulate
tensorboardX
uvicorn
dataclasses; python_version < '3.7'
starlette

# Requirements for running tests
blist; platform_system != "Windows"
azure-common
azure-mgmt-resource
msrestazure
boto3
cython >= 0.29.15
dataclasses; python_version < '3.7'
dask[complete]
feather-format
google-api-python-client
gym
gym-minigrid
kubernetes
lxml
modin
mock
moto
mypy
networkx
numba
# higher version of llvmlite breaks windows
llvmlite==0.34.0
openpyxl
opentelemetry-api==1.1.0
opentelemetry-sdk==1.1.0
opentelemetry-exporter-otlp==1.1.0
pexpect
Pillow; platform_system != "Windows"
pygments
pytest==6.2.5
pytest-asyncio
pytest-rerunfailures
pytest-sugar
pytest-timeout
scikit-learn==0.22.2
tensorflow
testfixtures==6.18.3
werkzeug
xlrd
starlette
fastapi
smart_open[s3]
tqdm
async-exit-stack
async-generator
# pyarrow v4.0.0 doesn't work in macOS
pyarrow==3.0.0
zdfs-dfs==2.1.1.2